Refactor: ADAPT-3750 Extract Navigation panel to setup/navigationPage.tsx + logo Change action
- Move the Navigation panel implementation out of SetupPage.tsx into a
  self-contained setup/navigationPage.tsx (exported as NavigationPage),
  mirroring setup/technicalSettingPage.tsx as agreed in review. SetupPage now
  imports and renders NavigationPage; unused imports removed. Pure move, no
  behaviour change (Navigation-only helpers NavAccordion/NavSelect/NavTextInput
  and constants moved; a local CheckboxRow copy avoids a circular import).
- Header Logo: add a 'Change' action next to 'Remove' so an existing logo can be
  swapped via the Asset Picker without removing it first.
